Revert "[FEATURE #17] Vehículos"
