+++ package.json
@@ -39,18 +39,14 @@
     "@nextui-org/tooltip": "^2.0.41",
     "@react-aria/visually-hidden": "3.8.12",
     "@react-pdf/renderer": "^3.4.4",
-    "apollo-upload-client": "^18.0.1",
     "axios": "^1.7.7",
     "clsx": "2.1.1",
     "framer-motion": "~11.1.1",
     "graphql": "^16.9.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-dropzone": "^14.2.9",
     "react-router-dom": "6.23.0",
     "react-select": "^5.8.0",
-    "react-slick": "^0.30.2",
-    "slick-carousel": "^1.8.1",
     "tailwind-variants": "0.1.20",
     "tailwindcss": "3.4.3",
     "uuid": "^10.0.0"
@@ -58,7 +54,6 @@
   "devDependencies": {
     "@types/react": "18.3.3",
     "@types/react-dom": "18.3.0",
-    "@types/react-slick": "^0.23.13",
     "@typescript-eslint/eslint-plugin": "^7.8.0",
     "@typescript-eslint/parser": "^7.8.0",
     "@vitejs/plugin-react": "^4.2.1",
