+++ src/components/TableHistorialLiquidaciones.tsx
@@ -383,7 +383,7 @@ export default function App() {
         <div className="flex justify-between gap-3 items-end">
           <Input
             isClearable
-            className="w-full sm:max-w-[24%]"
+            className="w-full sm:max-w-[44%]"
             placeholder="Buscar por conductor..."
             startContent={<SearchIcon />}
             value={filterValue}
