+++ src/graphql/anticipo.ts
@@ -1,0 +1,18 @@
+import { gql } from "@apollo/client";
+
+export const CREAR_ANTICIPOS = gql`
+    mutation RegistrarAnticipos($anticipos: [AnticipoInput!]!) {
+        registrarAnticipos(anticipos: $anticipos) {
+            id
+            valor
+            liquidacionId
+        }
+    }
+`;
+
+
+export const ELIMINAR_ANTICIPO = gql`
+  mutation EliminarAnticipo($id: ID!) {
+    eliminarAnticipo(id: $id)
+  }
+`;
