+++ src/graphql/conductor.ts
@@ -1,0 +1,13 @@
+import { gql } from "@apollo/client";
+
+export const OBTENER_CONDUCTORES = gql`
+    query ObtenerConductores {
+        obtenerConductores {
+            id
+            nombre
+            apellido
+            cc
+            salarioBase
+        }
+    }
+`
