+++ src/graphql/empresa.ts
@@ -1,0 +1,11 @@
+import { gql } from "@apollo/client";
+
+export const OBTERNER_EMPRESAS = gql`
+    query ObtenerEmpresas {
+        obtenerEmpresas {
+            id
+            NIT
+            Nombre
+        }
+    }
+`
