+++ src/graphql/liquidacion.ts
@@ -299,69 +299,3 @@ export const EDITAR_LIQUIDACION = gql`
 }
 `;
 
-export const OBTENER_VEHICULOS = gql`
-    query ObtenerVehiculos {
-        obtenerVehiculos {
-            id
-            placa
-        }
-    }
-` 
-
-export const OBTERNER_EMPRESAS = gql`
-    query ObtenerEmpresas {
-        obtenerEmpresas {
-            id
-            NIT
-            Nombre
-        }
-    }
-`
-
-export const OBTENER_CONDUCTORES = gql`
-    query ObtenerConductores {
-        obtenerConductores {
-            id
-            nombre
-            apellido
-            cc
-            salarioBase
-        }
-    }
-`
-
-export const OBTENER_CONFIGURACION_LIQUIDACION = gql`
-    query ConfiguracionesLiquidador {
-        configuracionesLiquidador {
-            id
-            nombre
-            valor
-        }
-    }
-
-`
-export const ACTUALIZAR_CONFIGURACION = gql`
-    mutation ActualizarConfiguracionesLiquidador($id: ID!, $input: ConfiguracionInput!) {
-        actualizarConfiguracionesLiquidador(id: $id, input: $input) {
-            id
-            nombre
-            valor
-        }
-    }
-`;
-
-export const CREAR_ANTICIPOS = gql`
-    mutation RegistrarAnticipos($anticipos: [AnticipoInput!]!) {
-        registrarAnticipos(anticipos: $anticipos) {
-            id
-            valor
-            liquidacionId
-        }
-    }
-`;
-
-export const ELIMINAR_ANTICIPO = gql`
-  mutation EliminarAnticipo($id: ID!) {
-    eliminarAnticipo(id: $id)
-  }
-`;
