+++ src/pages/liquidador.tsx
@@ -7,7 +7,7 @@ import { Tabs, Tab } from "@nextui-org/tabs";
 
 export default function Liquidador() {
   return (
-    <div className="md:px-20 space-y-8">
+    <div className="space-y-8">
       <div className="flex flex-col space-y-2">
         <Tabs className="mx-auto" color="primary" aria-label="Sections">
           <Tab key="liquidar" title="Liquidar">
