+++ tailwind.config.js
@@ -10,11 +10,7 @@ module.exports = {
     './node_modules/@nextui-org/theme/dist/**/*.{js,ts,jsx,tsx}',
   ],
   theme: {
-    extend: {
-      maxHeight: {
-        '1/3': '2%', // Un tercio de la altura
-      },
-    },
+    extend: {},
   },
   darkMode: "class",
   plugins: [nextui()],
